footer bugfix
